requirements.txt
# Core dependencies
pydantic>=2.6.0
requests>=2.31.0
playwright>=1.41.0

# LLM and Extraction
groq>=0.3.0
ollama>=0.1.0
litellm>=1.20.0

# Development and Testing
pytest>=7.4.0
mypy>=1.8.0
black>=24.1.0
ruff>=0.2.0

# Logging and Configuration
python-dotenv>=1.0.0
